services/storage-account/func-get-subscription-list/requirements.txt
azure-functions
azure-mgmt-subscription==3.1.1
azure-mgmt-resource
azure.identity==1.14.1
azure.data.tables==12.4.0
azure-keyvault-secrets
black==23.11.0
pandas==2.1.1
pytest==7.4.2
pytest-cov
python-dotenv==1.0.0